edit user info permisson

+++ app/Http/Controllers/UserController.php
@@ -22,27 +22,27 @@ class UserController extends Controller
      * Show the form for editing the specified resource.
      */
     public function edit(User $user)
-
     {
+        $this->authorize('update', $user);
         $editing = true;
         $ideas = $user->ideas()->paginate(5);
         return view('users.edit', compact('user', 'editing', 'ideas'));
     }
-
     /**
      * Update the specified resource in storage.
      */
     public function update(User $user)
     {
-        //dd(request()->all());
+
+        $this->authorize('update', $user);
         $validated = request()->validate([
             'name' => 'required|min:3|max:40',
             'bio' => 'nullable|min:1|max:255',
             'image' => 'image',
         ]);
 
-        if (request()->has('image')){
-            $imagePath = request()->file('image')->store('profile','public');
+        if (request()->has('image')) {
+            $imagePath = request()->file('image')->store('profile', 'public');
             $validated['image'] = $imagePath;
 
             Storage::disk('public')->delete($user->image ?? '');
@@ -50,7 +50,6 @@ class UserController extends Controller
 
         $user->update($validated);
         return redirect()->route('profile');
-
     }
 
     public function profile()
